Rename bounds modules.
