+++ src/optimagic/parameters/bounds.py
