+++ tests/optimagic/parameters/test_bounds.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 import pytest
 from optimagic.exceptions import InvalidBoundsError
-from optimagic.parameters.parameter_bounds import get_bounds
+from optimagic.parameters.bounds import get_bounds
 from numpy.testing import assert_array_equal
